break exception out to own class
