+++ modules/extension/graph/src/main/java/org/geotools/graph/path/WrongPathException.java
@@ -1,0 +1,15 @@
+package org.geotools.graph.path;
+
+@SuppressWarnings("serial")
+class WrongPathException extends Exception {
+    String message;
+
+    public WrongPathException(String msj) {
+        message = msj;
+    }
+
+    @Override
+    public String getMessage() {
+        return message;
+    }
+}
